Refactor CameraCapture to use more performant version
Also re-enable raster prop monitor cameras
