+++ Telemachus/src/CameraSnapshots/CameraCapture.cs
@@ -2,6 +2,7 @@
 using System.Collections;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Telemachus.CameraSnapshots
 {
@@ -23,8 +24,23 @@ namespace Telemachus.CameraSnapshots
         }
 
         protected Dictionary<string, Camera> cameraDuplicates = new Dictionary<string, Camera>();
-        protected List<string> activeCameras;
+        protected List<string> activeCameras = new List<string>();
         protected static readonly string[] skippedCameras = { "UIMainCamera", "UIVectorCamera", "velocity camera" };
+        private readonly string[] knownCameraNames =
+        {
+            "GalaxyCamera",
+            "Camera ScaledSpace",
+            "Camera VE Underlay", // Environmental Visual Enhancements plugin camera
+            "Camera VE Overlay",  // Environmental Visual Enhancements plugin camera
+            "Camera 01",
+            "Camera 00",
+            "InternalCamera",
+            "FXCamera"
+        };
+
+        public Dictionary<string, Camera> gameCameraMapping = new Dictionary<string, Camera>();
+
+
         protected string cameraContainerNamePrefix {
             get
             {
@@ -38,23 +54,142 @@ namespace Telemachus.CameraSnapshots
 
         protected virtual void LateUpdate()
         {
-            //PluginLogger.debug("LATEUPDATE FOR CAMERA");
-            if(CameraManager.Instance != null && HighLogic.LoadedSceneIsFlight)
+            //PluginLogger.debug("LATE UPDATE FOR FLIGHT CAMERA");
+            if (CameraManager.Instance != null && HighLogic.LoadedSceneIsFlight)
             {
-                //PluginLogger.debug("CAMERA FOUND, TAKING SCREENSHOT");
-                StartCoroutine(NewScreenshot());
+                duplicateAnyNewCameras();
+                repositionCamera();
+                StartCoroutine(newRender());
             }
         }
 
-        public void UpdateCameras()
+        public IEnumerator newRender()
         {
-            PluginLogger.debug("UPDATING CAMERAS");
+            yield return true;
+            //PluginLogger.debug(cameraManagerName() +  ": WAITING FOR END OF FRAME");
+            //yield return new WaitForEndOfFrame();
+            //PluginLogger.debug(cameraManagerName() + ": OUT OF FRAME");
+            Texture2D texture = getTexture2DFromRenderTexture();
+            this.imageBytes = texture.EncodeToJPG();
+            this.didRender = true;
+
+            Destroy(texture);
+        }
+
+        public Texture2D getTexture2DFromRenderTexture()
+        {
+            Texture2D texture2D = new Texture2D(overviewTexture.width, overviewTexture.height);
+            RenderTexture.active = overviewTexture;
+            texture2D.ReadPixels(new Rect(0, 0, overviewTexture.width, overviewTexture.height), 0, 0);
+            texture2D.Apply();
+            return texture2D;
+        }
+
+        public void duplicateAnyNewCameras()
+        {
+            if (overviewTexture == null)
+            {
+                overviewTexture = new RenderTexture(cameraResolution, cameraResolution, 24);
+            }
+
+            List<string> currentlyActiveCameras = new List<string>();
+
+            foreach (Camera camera in Camera.allCameras)
+            {
+                
+                // Don't duplicate any cameras we're going to skip
+                if (skippedCameras.IndexOf(camera.name) != -1)
+                {
+                    continue;
+                }
+
+                // Don't duplicate cameras we don't know about
+                if (knownCameraNames.IndexOf(camera.name) == -1)
+                {
+                    continue;
+                }
+
+
+                if (!cameraDuplicates.ContainsKey(camera.name))
+                {
+                    var cameraDuplicateGameObject = new GameObject(cameraContainerNamePrefix + camera.name);
+                    Camera cameraDuplicate = cameraDuplicateGameObject.AddComponent<Camera>();
+                    cameraDuplicates[camera.name] = cameraDuplicate;
+                    cameraDuplicate.CopyFrom(camera);
+                    cameraDuplicate.fieldOfView = fovAngle;
+                    cameraDuplicate.aspect = aspect;
+
+                    cameraDuplicate.targetTexture = this.overviewTexture;
+                    if (camera.name == "Camera 00" || camera.name == "FXCamera")
+                    {
+                        cameraDuplicate.nearClipPlane = cameraDuplicate.farClipPlane / 8192.0f;
+                    }
+
+                    //Now that the camera has been duplicated, add it to the list of active cameras
+                    activeCameras.Add(camera.name);
+                    if (!gameCameraMapping.ContainsKey(camera.name))
+                    {
+                        gameCameraMapping[camera.name] = camera;
+                    }
+                }
+
+                //Mark that the camera is currently active
+                currentlyActiveCameras.Add(camera.name);
+            }
+
+            if (currentlyActiveCameras.Count > 0 && activeCameras.Count > 0)
+            {
+                IEnumerable<string> disabledCameras = activeCameras.Except(currentlyActiveCameras);
+                foreach (string disabledCamera in disabledCameras)
+                {
+                    if (cameraDuplicates.ContainsKey(disabledCamera))
+                    {
+                        Destroy(cameraDuplicates[disabledCamera]);
+                        cameraDuplicates.Remove(disabledCamera);
+                        activeCameras.Remove(disabledCamera);
+                    }
+                }
+            }
+        }
+
+
+        public virtual void repositionCamera()
+        {
+           foreach (KeyValuePair<string, Camera> KVP in cameraDuplicates)
+            {
+                Camera cameraDuplicate = KVP.Value;
+                Camera gameCamera = gameCameraMapping[KVP.Key];
+
+                cameraDuplicate.transform.position = gameCamera.transform.position;
+                cameraDuplicate.transform.rotation = gameCamera.transform.rotation;
+
+                additionalCameraUpdates(cameraDuplicate, gameCamera);
+            }
+        }
+
+        public void verboseCameraDebug(Camera camera)
+        {
+            string[] debugProperties = {
+                        "CAMERA INFO: " + camera.name,
+                        "TARGET DISPLAY: " + camera.targetDisplay,
+                        "TARGET TEXTURE: " + camera.targetTexture,
+                        "RENDERING PATH: " + camera.renderingPath,
+                        "ACTUAL RENDER PATH: " + camera.actualRenderingPath,
+                        "CAMERA TYPE: " + camera.cameraType,
+                        "GAME OBJECT: " + camera.gameObject
+                    };
+            PluginLogger.debug(String.Join("\n", debugProperties));
+        }
+
+        /*public void UpdateCameras()
+        {
             if (CameraManager.Instance != null)
             {
                 //PluginLogger.debug("CURRENT CAMERA MODE: " + CameraManager.Instance.currentCameraMode);
             }
 
             activeCameras = new List<string>();
+            PluginLogger.debug("UPDATING CAMERAS");
             foreach (Camera camera in Camera.allCameras)
             {
                 // debugCameraDetails(camera);
@@ -93,9 +228,9 @@ namespace Telemachus.CameraSnapshots
                 //Now that the camera has been duplicated, add it to the list of active cameras
                 activeCameras.Add(camera.name);
             }
-        }
+        }*/
 
-        public virtual void additionalCameraUpdates(Camera cam){ }
+        public virtual void additionalCameraUpdates(Camera dupliateCam, Camera gameCamera){ }
 
         public virtual void debugCameraDetails(Camera cam)
         {
@@ -104,7 +239,7 @@ namespace Telemachus.CameraSnapshots
 
         public virtual void BeforeRenderNewScreenshot(){ }
         
-        public IEnumerator NewScreenshot()
+        /*public IEnumerator NewScreenshot()
         {
             if (mutex)
             {
@@ -127,7 +262,7 @@ namespace Telemachus.CameraSnapshots
             this.imageBytes = SnapshotRenderer.renderSnaphot(renderingCameras, cameraResolution, cameraResolution);
             this.didRender = true;
             mutex = false;
-        }
+        }*/
 
     }
 }
